Refactor to use RenderedGameObject
No need for Empty now! GameObject is no longer an abstract class, so it
can be instantiated directly and won't be buffered to the graphics card
or drawn in any way.

This also paves the way for lights, since they can use the
customInitGameObject() method to send the appropriate data to the scene
when they are initialized.